Move zstack into core and rename to zmx
